- Entity id is now int instead of long, entity id pooling will be implemented that will reuse entity ids, and it is unlikely we will ever handle more than a few hundred thousand entities at once - got rid of World.addEntity, created entities are part of the world by default - Worldmade World.matches deprecated, since EntityManager.matches is deprecated as well - EntityManager.matches now calls getEntitiesFor(aspect) and as a result slow as hell (might be a good idea to modify it a little bit) - JOGLRenderer is using Point instead of Circle to batch draw the stars - added RotationSystem

+++ src/main/java/frigo/asteroids/core/Entity.java
@@ -3,9 +3,9 @@ package frigo.asteroids.core;
 
 public final class Entity extends Value {
 
-    public final long id;
+    public final int id;
 
-    Entity (long id) {
+    Entity (int id) {
         this.id = id;
     }
